Populate-Implementations/MySQL-Postgres-Populate/requirements.txt
mysql-connector-python==8.1.0
protobuf==4.21.12
tqdm==4.65.0
psycopg2==2.9.6
# psycog?